requirement.txt
aiohttp==3.9.4
aioprometheus==23.12.0
apex==0.1
autopep8==2.3.1
cvcuda==0.2.0
deepspeed==0.15.2
einops==0.8.0
faiss==1.5.3
fastapi==0.115.2
file_utils==0.0.1
filelock==3.16.1
fire==0.7.0
flash_attn==2.4.2
Flask==3.0.3
flask_restful==0.3.10
ftfy==6.3.0
grouped_gemm==0.1.6
huggingface_hub==0.23.5
icetk==0.0.7
langdetect==1.0.9
lm_dataformat==0.0.20
lm_eval==0.4.5
lsh==0.1.2
matplotlib==3.9.2
mmcv==2.2.0
modelscope==1.18.1
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
Pillow==10.4.0
prettytable==3.11.0
psutil==5.9.8
pyarrow==15.0.2
pydantic==2.9.2
pytest==8.3.3
pytest_mock==3.14.0
PyYAML==6.0.1
PyYAML==6.0.2
ray==2.37.0
regex==2024.5.15
Requests==2.32.3
rjieba==0.1.11
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.1.99
setuptools==71.1.0
six==1.16.0
spacy==3.8.2
tensorboard==2.11.2
tiktoken==0.6.0
timm==1.0.9
tldextract==5.1.2
torchvision==0.16.2
tqdm==4.66.4
transformers==4.43.1
uvicorn==0.31.1
wandb==0.18.3
webdataset==0.2.100
wrapt==1.16.0
xformers==0.0.28.post1
ipython
vllm==0.5.0.post1
scikit-learn
